microservices/tts/src/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
boto3==1.34.10
kafka-python==2.0.2
gunicorn==21.2.0
python-dotenv==1.0.0
